+++ max_div/_cli/bm_solver/_benchmark_solver_run.py
@@ -1,6 +1,7 @@
 from max_div.benchmarks import BenchmarkProblemFactory
 
 from .benchmark_initialization import benchmark_initialization_strategies
+from .benchmark_optimization import benchmark_optimization_strategies
 
 
 def run_solver_benchmark(name: str, markdown: bool, file: bool = False, speed: float = 0.0):
@@ -11,3 +12,4 @@ def run_solver_benchmark(name: str, markdown: bool, file: bool = False, speed: f
             run_solver_benchmark(problem_name, markdown, file, speed)
     else:
         benchmark_initialization_strategies(name, markdown, file, speed)
+        benchmark_optimization_strategies(name, markdown, file, speed)